requirements.txt
streamlit==1.41.1
opencv-python==4.10.0.84
Pillow==10.4.0

